oauth2ClientPy3/requirements.txt
basicauth==0.2
defusedxml==0.4.1
Django==1.8.2
oauthlib==0.7.2
PyJWT==1.3.0
python-social-auth==0.2.9
python3-openid==3.0.5
rauth==0.7.1
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
wheel==0.24.0
